refactor: 2단계 피드백 반영 - 테스트코드 중 중복되는 코드 abstract 클래스화 및 상속 로직으로 변경
